+++ src/app/activity-logs/layout.jsx
@@ -1,0 +1,7 @@
+export const metadata = {
+  title: 'Activity Logs',
+}
+
+export default function ActivityLogsLayout({ children }) {
+  return children
+}
