+++ src/app/dashboard/layout.jsx
@@ -1,0 +1,7 @@
+export const metadata = {
+  title: 'Dashboard',
+}
+
+export default function DashboardLayout({ children }) {
+  return children
+}
