+++ src/app/login/layout.jsx
@@ -1,0 +1,7 @@
+export const metadata = {
+  title: 'Super Admin Login',
+}
+
+export default function LoginLayout({ children }) {
+  return children
+}
