+++ src/app/settings/layout.jsx
@@ -1,0 +1,7 @@
+export const metadata = {
+  title: 'Settings',
+}
+
+export default function SettingsLayout({ children }) {
+  return children
+}
