+++ src/app/users/[userId]/layout.jsx
@@ -1,0 +1,7 @@
+export const metadata = {
+  title: 'User Profile',
+}
+
+export default function UserDetailLayout({ children }) {
+  return children
+}
