+++ src/app/users/layout.jsx
@@ -1,0 +1,7 @@
+export const metadata = {
+  title: 'Users',
+}
+
+export default function UsersLayout({ children }) {
+  return children
+}
