+++ tsconfig.json
@@ -6,6 +6,17 @@
       "dom.iterable",
       "esnext"
     ],
+      "baseUrl": "src",
+      "paths": {
+        "@/components/*": ["components/*"],
+        "@/pages/*": ["pages/*"],
+        "@/utils/*": ["utils/*"],
+        "@/middleware/*": ["lib/middleware/*"],
+        "@/api/*": ["lib/api/*"],
+        "@/styles/*": ["styles/*"],
+        "@/supabase/*": ["lib/supabase/*"],
+        "@/lib/*": ["lib/*"]
+      },
     "allowJs": true,
     "skipLibCheck": true,
     "strict": false,
